+++ packages/gcp-gcs/src/index.ts
@@ -1,7 +1,7 @@
 import { Storage, StorageOptions } from '@google-cloud/storage'
-import * as Effect from '@effect/io/Effect'
-import * as Layer from '@effect/io/Layer'
-import * as Context from '@effect/data/Context'
+import * as Effect from 'effect/Effect'
+import * as Layer from 'effect/Layer'
+import * as Context from 'effect/Context'
 
 export type GCS = Storage
 /**
@@ -11,6 +11,12 @@ export const GCS = Context.Tag<GCS>('@google-cloud/storage')
 
 type GCSWriteError = {
   _tag: 'GCSWriteError'
+  meassage: string
+  stack: unknown
+}
+
+type GCSUrlSigningError = {
+  _tag: 'GCSUrlSigningError'
   meassage: string
   stack: unknown
 }
@@ -31,6 +37,30 @@ export const write = (
       try: () => gcs.bucket(bucket).file(key).save(data),
       catch: (e) => ({
         _tag: 'GCSWriteError',
+        message: `${e}`,
+        error: (e as Error).stack,
+      }),
+    })
+  )
+
+export const getPresignedUrl = (
+  bucket: string,
+  key: string,
+  lifetime: number // time to live in milliseconds (relative to when URL is created)
+) =>
+  Effect.flatMap(GCS, (gcs) =>
+    Effect.tryPromise({
+      try: () =>
+        gcs
+          .bucket(bucket)
+          .file(key)
+          .getSignedUrl({
+            version: 'v4',
+            action: 'read',
+            expires: Date.now() + lifetime,
+          }),
+      catch: (e) => ({
+        _tag: 'GCSUrlSigningError',
         message: `${e}`,
         error: (e as Error).stack,
       }),