+++ packages/gcp-logging/src/index.ts
@@ -1,12 +1,15 @@
-import { pipe } from '@effect/data/Function'
-import * as Effect from '@effect/io/Effect'
-import * as FiberRef from '@effect/io/FiberRef'
-import * as FiberRefs from '@effect/io/FiberRefs'
-import * as Logger from '@effect/io/Logger'
-import * as Option from '@effect/data/Option'
-import * as List from '@effect/data/List'
-import * as HashMap from '@effect/data/HashMap'
-import * as Cause from '@effect/io/Cause'
+import { pipe } from 'effect/Function'
+import * as Effect from 'effect/Effect'
+import * as FiberRef from 'effect/FiberRef'
+import * as FiberRefs from 'effect/FiberRefs'
+import * as Logger from 'effect/Logger'
+import * as Option from 'effect/Option'
+import * as List from 'effect/List'
+import * as HashMap from 'effect/HashMap'
+import * as Cause from 'effect/Cause'
+import * as LogLevel from 'effect/LogLevel'
+import * as S from '@effect/schema/Schema'
+import * as Layer from 'effect/Layer'
 
 type LogMeta = Record<string, string>
 
@@ -30,7 +33,9 @@ export const logError = (message: string, data: LogMeta) =>
 export const logFatal = (message: string, data: LogMeta) =>
   Effect.locally(logMeta, data)(Effect.logFatal(message))
 
-export const customLogger = (logFn: (a: any) => void = console.log) =>
+const defaultLogFunction = (a: any) => console.log(JSON.stringify(a))
+
+export const customLogger = (logFn: (a: any) => void = defaultLogFunction) =>
   Logger.make<unknown, void>(
     ({
       fiberId,
@@ -82,3 +87,22 @@ export const withTrace =
       Effect.annotateLogs(GCP_LOG_TRACE_KEY, args.trace),
       Effect.annotateLogs(GCP_LOG_SPAN_KEY, args.span)
     )
+
+export const LogLevelSchema = S.literal<
+  [
+    LogLevel.All['_tag'],
+    LogLevel.Fatal['_tag'],
+    LogLevel.Error['_tag'],
+    LogLevel.Warning['_tag'],
+    LogLevel.Info['_tag'],
+    LogLevel.Debug['_tag'],
+    LogLevel.Trace['_tag'],
+    LogLevel.None['_tag']
+  ]
+>('All', 'Fatal', 'Error', 'Warning', 'Info', 'Debug', 'Trace', 'None')
+
+export const LogLayer = (level: LogLevel.Literal) =>
+  Layer.provideMerge(
+    Logger.replace(Logger.defaultLogger, customLogger()),
+    Logger.minimumLogLevel(LogLevel.fromLiteral(level))
+  )