+++ packages/stripe/jest.config.ts
@@ -1,0 +1,8 @@
+const config = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/test/**/*.ts'],
+  passWithNoTests: true,
+}
+
+export default config