+++ packages/gcp-gcs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @effect-use/gcp-gcs
 
+## 1.0.0
+
+### Major Changes
+
+- 38f0e5b: use effect package, upgrade schema
+
+## 0.2.0
+
+### Minor Changes
+
+- d6138f3: add getPresignedUrl method
+
+## 0.1.0
+
+### Minor Changes
+
+- 8fe9f3e: latest effect packages
+
 ## 0.0.1
 
 ### Patch Changes