+++ packages/gcp-logging/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @effect-use/gcp-logging
 
+## 1.0.0
+
+### Major Changes
+
+- 38f0e5b: use effect package, upgrade schema
+
+### Minor Changes
+
+- 470ce70: Add LogLayer to replace the default logger
+
+## 0.2.0
+
+### Minor Changes
+
+- 8fe9f3e: latest effect packages
+
+## 0.1.0
+
+### Minor Changes
+
+- bb6fe94: The default formatter will stringify the messages so the log output is on a single line
+
 ## 0.0.1
 
 ### Patch Changes