+++ packages/temporal-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @effect-use/temporal-config
 
+## 1.0.0
+
+### Major Changes
+
+- 38f0e5b: use effect package, upgrade schema
+
+## 0.1.0
+
+### Minor Changes
+
+- 8fe9f3e: latest effect packages
+
 ## 0.0.1
 
 ### Patch Changes